Move methods to ExpressionResultTransformer, add documentation
